+++ ambition_utils/rrule/models.py
@@ -1,15 +1,14 @@
 from __future__ import annotations
 from datetime import datetime, timedelta
 from dateutil import parser
-from dateutil.rrule import rrule
+from dateutil.rrule import rrule, rruleset
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.postgres.fields import JSONField
 from django.db import models, transaction
 from django.utils.module_loading import import_string
 from fleming import fleming
 from manager_utils import bulk_update
-from timezone_field import TimeZoneField
+from ambition_utils.fields import TimeZoneField
 from typing import List
 import copy
 import pytz
@@ -107,10 +106,13 @@ class RRule(models.Model):
     """
 
     # Params used to generate the rrule
-    rrule_params = JSONField()
+    rrule_params = models.JSONField()
+
+    # Optional params used to generate the rrule exclusion
+    rrule_exclusion_params = models.JSONField(default=None, blank=True, null=True)
 
     # Any meta data associated with the object that created this rule
-    meta_data = JSONField(default=dict)
+    meta_data = models.JSONField(default=dict)
 
     # The timezone all dates should be converted to
     time_zone = TimeZoneField(default='UTC')
@@ -133,10 +135,16 @@ class RRule(models.Model):
     # The name of the method to call on the related_object when the recurrence has expired
     related_object_handler_name = models.TextField(default=None, null=True, blank=True)
 
+    # An optional number of days to offset occurrences by
+    day_offset = models.SmallIntegerField(blank=True, null=True)
+
     # Custom object manager
     objects = RRuleManager()
 
     def get_time_zone_object(self):
+        """
+        Returns the time zone object from pytz
+        """
         if self.time_zone is None:
             return pytz.utc
 
@@ -158,13 +166,42 @@ class RRule(models.Model):
         except:
             return None
 
+    def get_rrule_set(self):
+        """
+        Returns the rrule set that will combine the rrule and optional exclusion rrule
+        """
+        rrule_set = rruleset()
+        rrule_set.rrule(self.get_rrule())
+        rrule_exclusion = self.get_rrule_exclusion()
+        if rrule_exclusion:
+            rrule_set.exrule(rrule_exclusion)
+        return rrule_set
+
     def get_rrule(self):
         """
         Builds the rrule object by restoring all the params.
+        """
+        return self.get_rrule_from_params(self.rrule_params)
+
+    def get_rrule_exclusion(self):
+        """
+        Builds the rrule exclusion object by restoring all the params.
+        :rtype: rrule
+        """
+        return self.get_rrule_from_params(self.rrule_exclusion_params)
+
+    def get_rrule_from_params(self, params):
+        """
+        Creates an rrule object from a dict of rrule params. Returns None if no params exists.
         The dtstart param will be converted to local time if it is set.
         :rtype: rrule
         """
-        params = copy.deepcopy(self.rrule_params)
+        # Check for none or empty
+        if not params:
+            return None
+
+        # Create a deep copy because we will manipulate
+        params = copy.deepcopy(params)
 
         # Convert next scheduled from utc back to time zone
         if params.get('dtstart') and not hasattr(params.get('dtstart'), 'date'):
@@ -180,24 +217,29 @@ class RRule(models.Model):
         # Return the rrule
         return rrule(**params)
 
-    def get_next_occurrence(self, last_occurrence=None, force=False):
+    def get_next_occurrence(self, last_occurrence=None, calculate_offset=True, force=False):
         """
         Builds the rrule and returns the next date in the series or None of it is the end of the series
         :param last_occurrence: The last occurrence that was generated
+        :param calculate_offset: Should the given offset be calculated?
         :param force: If the next occurrence is none, force the rrule to generate another
         :rtype: rrule or None
         """
         # Get the last occurrence
         last_occurrence = last_occurrence or self.last_occurrence or datetime.utcnow()
 
-        # Get the rule
-        rule = self.get_rrule()
+        # Get the rule set
+        rule_set = self.get_rrule_set()
 
         # Convert to local time zone for getting next occurrence, otherwise time zones ahead of utc will return the same
         last_occurrence = fleming.convert_to_tz(last_occurrence, self.get_time_zone_object(), return_naive=True)
 
+        # Un-offset the last occurrence to match the rule_set's dates for .after() before offsetting again later
+        if calculate_offset:
+            last_occurrence = self.offset(last_occurrence, reverse=True)
+
         # Generate the next occurrence
-        next_occurrence = rule.after(last_occurrence)
+        next_occurrence = rule_set.after(last_occurrence)
 
         # If next occurrence is none and force is true, force the rrule to generate another date
         if next_occurrence is None and force:
@@ -209,11 +251,11 @@ class RRule(models.Model):
             self.rrule_params.pop('count', None)
             self.rrule_params.pop('until', None)
 
-            # Refetch the rule
-            rule = self.get_rrule()
+            # Refetch the rule set
+            rule_set = self.get_rrule_set()
 
             # Generate the next occurrence
-            next_occurrence = rule.after(last_occurrence)
+            next_occurrence = rule_set.after(last_occurrence)
 
             # Restore the rrule params
             self.rrule_params = original_rrule_params
@@ -221,6 +263,9 @@ class RRule(models.Model):
         # If there is a next occurrence, convert to utc
         if next_occurrence:
             next_occurrence = self.convert_to_utc(next_occurrence)
+
+            if calculate_offset:
+                next_occurrence = self.offset(next_occurrence)
 
         # Return the next occurrence
         return next_occurrence
@@ -242,6 +287,7 @@ class RRule(models.Model):
             return False
 
         self.last_occurrence = self.next_occurrence
+
         self.next_occurrence = self.get_next_occurrence(self.last_occurrence)
 
         # Only save if the flag is true
@@ -261,6 +307,23 @@ class RRule(models.Model):
 
         return dt
 
+    def offset(self, dt, reverse=False) -> datetime:
+        """
+        Offsets a given datetime by the number of days specified by day_offset.
+        :param dt: The datetime to offset by day_offset
+        :param reverse: Reverse the offset calculation.
+        :return dt:
+        """
+        # The offset gets multiplied by 1 or -1 depending on offset direction
+        multiplier = -1 if reverse else 1
+
+        # Timezone is considered when not reversing offset for comparisons in rrule.after().
+        return fleming.add_timedelta(
+            dt,
+            timedelta(days=self.day_offset * multiplier),
+            within_tz=self.time_zone if not reverse else None
+        ) if self.day_offset else dt
+
     def refresh_next_occurrence(self, current_time=None):
         """
         Sets the next occurrence date based on the current rrule param definition. The date will be after the
@@ -274,9 +337,10 @@ class RRule(models.Model):
         next_occurrence = self.get_next_occurrence(last_occurrence=current_time)
 
         if next_occurrence:
-            # Only set if the new time is still greater than now
+            # Only set if the new time is still greater than now.
+            # Offset date if applicable.
             if next_occurrence > datetime.utcnow():
-                self.next_occurrence = next_occurrence
+                self.next_occurrence = self.offset(next_occurrence)
         else:
             self.next_occurrence = next_occurrence
 
@@ -284,28 +348,50 @@ class RRule(models.Model):
         self.set_date_objects()
 
     def set_date_objects(self):
+        """
+        Ensure that all the date keys are properly set on all rrule params
+        """
+
+        # Convert the rrule and exclusion rrule params to properly set date keys
+        is_new = self.pk is None
+        self.set_date_objects_for_params(self.rrule_params, is_new=is_new)
+        self.set_date_objects_for_params(self.rrule_exclusion_params, is_new=is_new)
+
         # Check if this is a new rrule object
-        if self.pk is None:
-            # Convert next scheduled from utc back to time zone
-            if self.rrule_params.get('dtstart') and not hasattr(self.rrule_params.get('dtstart'), 'date'):
-                self.rrule_params['dtstart'] = parser.parse(self.rrule_params['dtstart'])
-
-            # Convert until date from utc back to time zone
-            if self.rrule_params.get('until') and not hasattr(self.rrule_params.get('until'), 'date'):
-                self.rrule_params['until'] = parser.parse(self.rrule_params['until'])
-
+        if is_new:
             # Get the first scheduled time according to the rrule (this converts from utc back to local time)
-            self.next_occurrence = self.get_rrule()[0]
+            self.next_occurrence = self.get_rrule_set()[0]
 
             # Convert back to utc before saving
             self.next_occurrence = self.convert_to_utc(self.next_occurrence)
 
-        # Serialize the datetime objects if they exist
-        if self.rrule_params.get('dtstart') and hasattr(self.rrule_params.get('dtstart'), 'date'):
-            self.rrule_params['dtstart'] = self.rrule_params['dtstart'].strftime('%Y-%m-%d %H:%M:%S')
+            # Offset, if applicable, for new objects.
+            self.next_occurrence = self.offset(self.next_occurrence)
 
-        if self.rrule_params.get('until') and hasattr(self.rrule_params.get('until'), 'date'):
-            self.rrule_params['until'] = self.rrule_params['until'].strftime('%Y-%m-%d %H:%M:%S')
+    def set_date_objects_for_params(self, params, is_new=False):
+        """
+        Give an rrule params object, ensure that the date keys are properly set and properly converted to strings
+        """
+        # Check for no params
+        if not params:
+            return params
+
+        # Check if this is a new rrule object
+        if is_new:
+            # Convert next scheduled from utc back to time zone
+            if params.get('dtstart') and not hasattr(params.get('dtstart'), 'date'):
+                params['dtstart'] = parser.parse(params['dtstart'])
+
+            # Convert until date from utc back to time zone
+            if params.get('until') and not hasattr(params.get('until'), 'date'):
+                params['until'] = parser.parse(params['until'])
+
+        # Serialize the datetime objects if they exist
+        if params.get('dtstart') and hasattr(params.get('dtstart'), 'date'):
+            params['dtstart'] = params['dtstart'].strftime('%Y-%m-%d %H:%M:%S')
+
+        if params.get('until') and hasattr(params.get('until'), 'date'):
+            params['until'] = params['until'].strftime('%Y-%m-%d %H:%M:%S')
 
     def save(self, *args, **kwargs):
         """
@@ -313,6 +399,8 @@ class RRule(models.Model):
         determined and set. The `dtstart` and `until` objects will be safely encoded as strings if they are
         datetime objects.
         """
+
+        # Run any pre save hooks
         self.pre_save_hooks()
 
         # Call the parent save method
@@ -325,34 +413,38 @@ class RRule(models.Model):
         :param start_date: The optional start date to begin generating dates after
         :return: A list of datetime objects
         """
+
+        # Assert that we have dates
         assert num_dates > 0
 
+        # Ensure that pre save hooks have been run
         self.pre_save_hooks()
 
+        # Generate the dates
         dates = []
-
         try:
             # Capture the rule's first date for use in RRule.after() in the loop.
-            rule = self.get_rrule()
+            rule_set = self.get_rrule_set()
 
             # Evaluate if the first date should be retained.
-            d = self.convert_to_utc(rule[0])
+            d = self.convert_to_utc(rule_set[0])
             if not start_date or d > start_date:
-                dates.append(d)
+                dates.append(self.offset(d))
 
             # Continue evaluating and appending dates to satisfy desired number,
             # retaining date for evaluation in the next iteration.
+            # The offset is ignored for this comparison and applied at appending.
             while len(dates) < num_dates:
-                d = self.get_next_occurrence(last_occurrence=d)
+                d = self.get_next_occurrence(last_occurrence=d, calculate_offset=False)
                 if d:
                     if not start_date or d > start_date:
-                        dates.append(d)
+                        dates.append(self.offset(d))
                 else:
                     break
-
         except Exception:  # pragma: no cover
             pass
 
+        # Return the generated dates
         return dates
 
     def generate_dates(self, num_dates=20):
@@ -371,48 +463,19 @@ class RRule(models.Model):
         # Clear id to force a new object.
         clone = copy.deepcopy(self)
         clone.id = None
-
         clone.save()
-
         return clone
 
     def clone_with_day_offset(self, day_offset: int) -> RRule:
         """
-        Creates a clone of a passed RRule object offset by a specified number of days
+        Creates a clone of a passed RRule object with day_offset set.
+        clone() is not called to ensure .id is not set before .save() so offset is applied.
         :param day_offset: The number of days to offset the clone's start date. Can be negative.
         """
-
-        # Create a clone of itself
-        clone = self.clone()
-
-        # Manually update the rrule.dtstart & next_occurrence with the offset.
-        clone.rrule_params['dtstart'] = parser.parse(clone.rrule_params['dtstart']) + timedelta(days=day_offset)
-        clone.next_occurrence = clone.next_occurrence + timedelta(days=day_offset)
-
-        # Update until param by offsetting if it exists
-        if 'until' in clone.rrule_params:
-            clone.rrule_params['until'] = parser.parse(clone.rrule_params['until']) + timedelta(days=day_offset)
-
-        def offset_day(day: int) -> int:
-            """
-            Calculates the representation of a given day of the week plus the provided offset
-            For example, Tuesday (1) - 3 days yields Saturday (5).
-            :param int day: 0-6 that corresponds to RRule's weekday constants, MO-SU.
-            """
-            return (7 + (day + day_offset)) % 7
-
-        # Update byweekday param by offsetting. byweekday can be an array or integer.
-        if 'byweekday' in clone.rrule_params:
-            if isinstance(clone.rrule_params['byweekday'], list):
-                clone.rrule_params['byweekday'] = [
-                    offset_day(day) for day in clone.rrule_params['byweekday']
-                ]
-            else:
-                clone.rrule_params['byweekday'] = offset_day(clone.rrule_params['byweekday'])
-
-        # Lock it.
+        clone = copy.deepcopy(self)
+        clone.id = None
+        clone.day_offset = day_offset
         clone.save()
-
         return clone
 
     @classmethod